+++ libsrc/ffdec_lib/test/com/jpexs/decompiler/flash/as3decompile/ActionScript3CrossCompileSwfToolsDecompileTest.java
@@ -521,9 +521,9 @@ public class ActionScript3CrossCompileSwfToolsDecompileTest extends ActionScript
 
     @Test
     public void testTryCatchWith() {
-        decompileMethod("swftools", "testTryCatchWith", "var _loc1_:* = new MyTest();\r\n"
+        decompileMethod("swftools", "testTryCatchWith", "var _loc1_:MyTest = new MyTest();\r\n"
                 + "trace(\"before with\");\r\n"
-                + "var _loc2_:* = _loc1_;\r\n"
+                + "var _loc2_:MyTest = _loc1_;\r\n"
                 + "with(_loc2_)\r\n"
                 + "{\r\n"
                 + "trace(\"before try\");\r\n"